+++ Dice/RustyDie.cs
@@ -34,19 +34,19 @@ namespace MysteryDice.Dice
                 switch (diceRoll)
                 {
                     case 3:
-                        Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(1, 2));
+                        Networker.Instance.JackpotServerRPC(PlayerUser.playerClientId, UnityEngine.Random.Range(1, 2));
                         Misc.SafeTipMessage($"Rolled 3", "Spawning some scrap");
                         break;
                     case 4:
-                        Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(3, 4));
+                        Networker.Instance.JackpotServerRPC(PlayerUser.playerClientId, UnityEngine.Random.Range(3, 4));
                         Misc.SafeTipMessage($"Rolled 4", "Spawning scrap");
                         break;
                     case 5:
-                        Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(5, 6));
+                        Networker.Instance.JackpotServerRPC(PlayerUser.playerClientId, UnityEngine.Random.Range(5, 6));
                         Misc.SafeTipMessage($"Rolled 5", "Spawning more scrap");
                         break;
                     case 6:
-                        Networker.Instance.JackpotServerRPC(PlayerUser.actualClientId, UnityEngine.Random.Range(7, 8));
+                        Networker.Instance.JackpotServerRPC(PlayerUser.playerClientId, UnityEngine.Random.Range(7, 8));
                         Misc.SafeTipMessage($"Rolled 6", "Spawning a lot of scrap!");
                         break;
                 }