+++ Effects/TarotCards.cs
@@ -1,52 +1,41 @@
 ﻿using GameNetcodeStuff;
-using LethalLib.Modules;
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using Unity.Netcode;
 using UnityEngine;
 
 namespace MysteryDice.Effects
 {
-    internal class ScrapJackpot : IEffect
+    internal class TarotCards : IEffect
     {
-        public string Name => "Scrap jackpot";
+        public string Name => "Tarot Cards";
         public EffectType Outcome => EffectType.Great;
         public bool ShowDefaultTooltip => false;
-        public string Tooltip => "Spawning more scrap!";
+        public string Tooltip => "A different kind of gambling!";
 
         public void Use()
         {
-            Networker.Instance.JackpotServerRPC(GameNetworkManager.Instance.localPlayerController.playerClientId, UnityEngine.Random.Range(3, 9));
+            Networker.Instance.TarotServerRPC(GameNetworkManager.Instance.localPlayerController.playerClientId);
         }
-
-
-
-        public static void JackpotScrap(ulong userID, int amount)
+        public static void TarotScrap(ulong userID)
         {
             PlayerControllerB player = Misc.GetPlayerByUserID(userID);
 
+            List<Item> items = UnityEngine.Resources.FindObjectsOfTypeAll<Item>().ToList();
+            Item TarotCardsItem = items.LastOrDefault(item => item.name.Equals("TarrotCard") && item.spawnPrefab != null);
             RoundManager RM = RoundManager.Instance;
             List<Item> scrapToSpawn = new List<Item>();
-
+            //tarrot_item_prefab(Clone) with prefab name tarrot_item_prefab, and properties name TarrotCard + Tarrot Card
             List<NetworkObjectReference> netObjs = new List<NetworkObjectReference>();
             List<int> scrapValues = new List<int>();
             List<float> scrapWeights = new List<float>();
 
-            int amountOfScrap = amount;
-
-            List<int> weightList = new List<int>(RM.currentLevel.spawnableScrap.Count);
-            for (int j = 0; j < RM.currentLevel.spawnableScrap.Count; j++)
-            {
-                weightList.Add(RM.currentLevel.spawnableScrap[j].rarity);
-            }
-            int[] weights = weightList.ToArray();
-
-            for (int i = 0; i < amountOfScrap; i++)
-            {
-                scrapToSpawn.Add(RM.currentLevel.spawnableScrap[RM.GetRandomWeightedIndex(weights)].spawnableItem);
-            }
-
+            scrapToSpawn.Add(TarotCardsItem);
+            
             foreach (Item scrap in scrapToSpawn)
             {
                 GameObject obj = UnityEngine.Object.Instantiate(scrap.spawnPrefab, player.transform.position, Quaternion.identity, RM.spawnedScrapContainer);
@@ -56,21 +45,19 @@ namespace MysteryDice.Effects
                 component.scrapValue = (int)(UnityEngine.Random.Range(scrap.minValue, scrap.maxValue) * RM.scrapValueMultiplier);
                 scrapValues.Add(component.scrapValue);
                 scrapWeights.Add(component.itemProperties.weight);
-
                 NetworkObject netObj = obj.GetComponent<NetworkObject>();
-                netObj.Spawn(); 
+                netObj.Spawn();
                 component.FallToGround(true);
                 netObjs.Add(netObj);
             }
 
             RM.StartCoroutine(DelayedSync(RM, netObjs.ToArray(), scrapValues.ToArray(), scrapWeights.ToArray()));
         }
-
         public static IEnumerator DelayedSync(RoundManager RM, NetworkObjectReference[] netObjs, int[] scrapValues, float[] scrapWeights)
         {
             yield return new WaitForSeconds(3f);
-            RM.SyncScrapValuesClientRpc(netObjs, scrapValues); // Sync to clients
-            Networker.Instance.SyncItemWeightsClientRPC(netObjs, scrapWeights); // Sync weights
+            RM.SyncScrapValuesClientRpc(netObjs, scrapValues);
+            Networker.Instance.SyncItemWeightsClientRPC(netObjs, scrapWeights);
         }
     }
 }