+++ Patches/GetEnemies.cs
@@ -23,8 +23,8 @@ namespace MysteryDice.Patches
     internal class GetEnemies
     {
 
-        public static SpawnableEnemyWithRarity Masked, HoardingBug, Centipede, Dog, Jester, Bracken, Stomper, Coilhead, Beehive, Sandworm, Spider, Giant, Maneater, Shrimp, CrystalRay, Lasso, Barber;
-        public static SpawnableMapObject SpawnableLandmine, SpawnableTurret, SpawnableTP, SpawnableSpikeTrap; 
+        public static SpawnableEnemyWithRarity Masked, HoardingBug, Centipede, Dog, Jester, Bracken, Stomper, Coilhead, Beehive, Sandworm, Spider, Giant, Maneater, Shrimp, CrystalRay, Lasso, Barber, BellCrab, Urchin, Horse, Nemo, Bruce, MantisShrimp;
+        public static SpawnableMapObject SpawnableLandmine, SpawnableTurret, SpawnableTP, SpawnableSpikeTrap, Seamine, Bertha; 
         private static readonly string teleporterTrapId = "TeleporterTrap"; 
 
         [HarmonyPatch("Start")]
@@ -35,7 +35,7 @@ namespace MysteryDice.Patches
             {
                 foreach (SpawnableEnemyWithRarity enemy in level.Enemies)
                 {
-                    //MysteryDice.CustomLogger.LogInfo("Enemy Found: " + enemy.enemyType.enemyName);
+                    //MysteryDice.CustomLogger.LogInfo("Enemy Found Inside: " + enemy.enemyType.enemyName);
                     if (enemy.enemyType.enemyName == "Masked")
                         Masked = enemy;
                     if (enemy.enemyType.enemyName == "Hoarding bug")
@@ -62,24 +62,36 @@ namespace MysteryDice.Patches
                         Lasso = enemy;
                     if (enemy.enemyType.enemyName == "Clay Surgeon")
                         Barber = enemy;
+                    if (enemy.enemyType.enemyName == "Mantis Shrimp")
+                        MantisShrimp = enemy;
+                    if (enemy.enemyType.enemyName == "BellCrab")
+                        BellCrab = enemy;
                 }
 
                 foreach (SpawnableEnemyWithRarity enemy in level.DaytimeEnemies)
                 {
-                    //MysteryDice.CustomLogger.LogInfo("Enemy Found: " + enemy.enemyType.enemyName);
+                    //MysteryDice.CustomLogger.LogInfo("Enemy Found Daytime: " + enemy.enemyType.enemyName);
                     if (enemy.enemyType.enemyName == "Red Locust Bees")
                         Beehive = enemy;
+                    if (enemy.enemyType.enemyName == "Urchin")
+                        Urchin = enemy;
+                    if (enemy.enemyType.enemyName == "Horse")
+                        Horse = enemy;
+                    if (enemy.enemyType.enemyName == "Nemo")
+                        Nemo = enemy;
                 }
 
                 foreach (SpawnableEnemyWithRarity enemy in level.OutsideEnemies)
                 {
-                    //MysteryDice.CustomLogger.LogInfo("Enemy Found: " + enemy.enemyType.enemyName);
+                    //MysteryDice.CustomLogger.LogInfo("Enemy Found Outside: " + enemy.enemyType.enemyName);
                     if (enemy.enemyType.enemyName == "Earth Leviathan") 
                         Sandworm = enemy;
                     if (enemy.enemyType.enemyName == "ForestGiant")
                         Giant = enemy;
                     if (enemy.enemyType.enemyName == "MouthDog")
                         Dog = enemy;
+                    if (enemy.enemyType.enemyName == "Bruce")
+                        Bruce = enemy;
                 }
 
                 foreach (var item in level.spawnableMapObjects)
@@ -96,6 +108,12 @@ namespace MysteryDice.Patches
 
                     if (item.prefabToSpawn.name == "TeleporterTrap" && SpawnableTP == null)
                         SpawnableTP = item;
+
+                    if (item.prefabToSpawn.name == "Seamine" && Seamine == null)
+                        Seamine = item;
+
+                    if (item.prefabToSpawn.name == "Bertha" && Bertha == null)
+                        Bertha = item;
 
                 }